+++ CHANGELOG.md
@@ -8,11 +8,15 @@
 - Fix errors in config import #882
 - Save current spyglass version in analysis nwb files to aid diagnosis #897
 - Add functionality to export vertical slice of database. #875
+- Add pynapple support #898
 
 ### Pipelines
 
--Spikesorting
+- Spikesorting
     - Update calls in v0 pipeline for spikeinterface>=0.99 #893
+    - Fix method type of `get_spike_times` #904
+- Decoding
+    - Handle dimensions of clusterless `get_ahead_behind_distance` #904
 
 ## [0.5.1] (March 7, 2024)