+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - Save current spyglass version in analysis nwb files to aid diagnosis #897
 - Add pynapple support #898
 - Prioritize datajoint filepath entry for defining abs_path of analysis nwbfile #918
+- Fix potential duplicate entries in Merge part tables #922
 
 ### Pipelines