+++ CHANGELOG.md
@@ -2,8 +2,32 @@
 
 ## [0.5.2] (Unreleased)
 
+### Release Notes
+
+<!-- Running draft to be removed immediately prior to release. -->
+
+### Infrastructure
+
 - Refactor `TableChain` to include `_searched` attribute. #867
 - Fix errors in config import #882
+- Save current spyglass version in analysis nwb files to aid diagnosis #897
+- Add functionality to export vertical slice of database. #875
+- Add pynapple support #898
+- Update PR template checklist to include db changes. #903
+- Avoid permission check on personnel tables. #903
+- Add documentation for `SpyglassMixin`. #903
+- Add helper to identify merge table by definition. #903
+- Prioritize datajoint filepath entry for defining abs_path of analysis nwbfile #918
+- Fix potential duplicate entries in Merge part tables #922
+
+### Pipelines
+
+- Spikesorting
+    - Update calls in v0 pipeline for spikeinterface>=0.99 #893
+    - Fix method type of `get_spike_times` #904
+- Decoding
+    - Handle dimensions of clusterless `get_ahead_behind_distance` #904
+    - Fix improper handling of nwb file names with .strip #929
 
 ## [0.5.1] (March 7, 2024)
 
@@ -22,7 +46,7 @@
     - Fixes to `_convert_mp4` #834
     - Replace deprecated calls to `yaml.safe_load()` #834
 - Spikesorting:
-    - Increase`spikeinterface` version to >=0.99.1, <0.100 #852
+    - Increase`spikeinterface` version to >=0.99.1, \<0.100 #852
     - Bug fix in single artifact interval edge case #859
     - Bug fix in FigURL #871
 - LFP
@@ -199,3 +223,5 @@
 [0.4.2]: https://github.com/LorenFrankLab/spyglass/releases/tag/0.4.2
 [0.4.3]: https://github.com/LorenFrankLab/spyglass/releases/tag/0.4.3
 [0.5.0]: https://github.com/LorenFrankLab/spyglass/releases/tag/0.5.0
+[0.5.1]: https://github.com/LorenFrankLab/spyglass/releases/tag/0.5.1
+[0.5.2]: https://github.com/LorenFrankLab/spyglass/releases/tag/0.5.2